+++ src/theticketbot/migrations/005-add-inbox-destination.sql
@@ -1,0 +1,3 @@
+ALTER TABLE inbox
+    ADD COLUMN
+        destination_id INTEGER REFERENCES channel (id) ON DELETE SET NULL;