+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### New Features
+
+- Add `/inbox destination` to allow routing tickets to any channel
+
+Localizations have not yet been updated.
+
 ## [0.4.3] - 2024-07-01
 
 ### New Features